# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/reservation": {
            "post": {
                "description": "Create a reservation with room details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation"
                ],
                "summary": "Create a new reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003cJWT Token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Reservation Information",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponseCreateReservation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/calculation": {
            "get": {
                "description": "Calculate the total cost for a room reservation based on room, snack, participant count, and time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation"
                ],
                "summary": "Endpoint for calculating reservation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "room_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Snack ID",
                        "name": "snack_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time (YYYY-MM-DD HH:MM:SS)",
                        "name": "startTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End time (YYYY-MM-DD HH:MM:SS)",
                        "name": "endTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of participants",
                        "name": "participant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Phone number",
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Company name",
                        "name": "company",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer \u003cJWT Token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponseReservationCalculation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/status/{id}": {
            "put": {
                "description": "Update the status of a reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation"
                ],
                "summary": "Update reservation status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reservation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer \u003cJWT Token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "New reservation status",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateReservationStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponseUpdateReservationStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/{id}": {
            "get": {
                "description": "Get the details of a reservation by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation"
                ],
                "summary": "Get reservation by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003cJWT Token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Reservation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ReservationCalculationReservation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.PersonalData": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "endTime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "participant": {
                    "type": "integer"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "models.ReservationCalculationData": {
            "type": "object",
            "properties": {
                "personalData": {
                    "$ref": "#/definitions/models.PersonalData"
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RoomCalculation"
                    }
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "models.ReservationCalculationReservation": {
            "type": "object",
            "properties": {
                "personalData": {
                    "$ref": "#/definitions/models.PersonalData"
                },
                "rooms": {
                    "description": "Ubah ke RoomCalculationReservation",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RoomCalculationReservation"
                    }
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "models.ReservationRequest": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RoomRequest"
                    }
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "models.RoomCalculation": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "imageURL": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pricePerHour": {
                    "type": "number"
                },
                "snack": {
                    "$ref": "#/definitions/models.SnackCategory"
                },
                "subTotalRoom": {
                    "type": "number"
                },
                "subTotalSnack": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.RoomCalculationReservation": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "duration": {
                    "description": "Duration in hours",
                    "type": "integer"
                },
                "endTime": {
                    "type": "string"
                },
                "imageURL": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "participant": {
                    "type": "integer"
                },
                "pricePerHour": {
                    "type": "number"
                },
                "snack": {
                    "$ref": "#/definitions/models.SnackCategory"
                },
                "startTime": {
                    "type": "string"
                },
                "subTotalRoom": {
                    "type": "number"
                },
                "subTotalSnack": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.RoomRequest": {
            "type": "object",
            "properties": {
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "participant": {
                    "type": "integer"
                },
                "snackID": {
                    "type": "integer"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
...
        "models.SnackCategory": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "models.SuccessResponseCreateReservation": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.SuccessResponseReservationCalculation": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.ReservationCalculationData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.SuccessResponseUpdateReservationStatus": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateReservationStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "e.g., \"confirmed\", \"cancelled\", etc.",
                    "type": "string"
                }
            }
        },
...
    }
}